+++ ksqlDb.RestApi.Client.ProtoBuf/ChangeLog.md
@@ -1,3 +1,4 @@
 ﻿# ksqlDB.RestApi.Client.ProtoBuf
 
 # v0.1.0
+- ProtoBufKSqlDbContext - adds support for Protobuf content type for query-stream endpoints